+++ Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Added
+- Added [BaryScore, InfoLM, and DepthScore](models/colombo2021/Readme.md)
+
 ### Changed
 - Changed the backend implementation of MoverScore to use a non-IDF dict based version.